+++ idunn/api/place/__init__.py
@@ -1,0 +1,5 @@
+from .address import Address
+from .admin import Admin
+from .poi import POI
+from .street import Street
+from .place import Place
