+++ idunn/api/place/admin.py
@@ -1,8 +1,8 @@
-from idunn.api.place import Place
-from idunn.utils.geom import build_blocks, get_geom
+from .place import Place
+from idunn.api.utils import build_blocks, get_geom
 
-class Street(Place):
-    PLACE_TYPE = 'street'
+class Admin(Place):
+    PLACE_TYPE = 'admin'
 
     @classmethod
     def load_place(cls, es_place, lang, settings, verbosity):
@@ -12,11 +12,10 @@ class Street(Place):
             id=es_place.get('id', ''),
             name=es_place.get('name', ''),
             local_name=es_place.get('name'),
-            class_name=es_place.get('poi_class'),
-            subclass_name=es_place.get('poi_subclass'),
+            class_name='city',
+            subclass_name='city',
             geometry=get_geom(es_place),
-            address= {
-                'label': address.get('label')
-            },
+            label=address.get('label'),
+            address=address,
             blocks=build_blocks(es_place, lang, verbosity)
         )
