+++ idunn/api/place/street.py
@@ -1,8 +1,8 @@
-from idunn.api.place import Place
-from idunn.utils.geom import build_blocks, get_geom
+from .place import Place
+from idunn.api.utils import build_blocks, get_geom
 
-class Address(Place):
-    PLACE_TYPE = 'address'
+class Street(Place):
+    PLACE_TYPE = 'street'
 
     @classmethod
     def load_place(cls, es_place, lang, settings, verbosity):
@@ -15,8 +15,7 @@ class Address(Place):
             class_name=es_place.get('poi_class'),
             subclass_name=es_place.get('poi_subclass'),
             geometry=get_geom(es_place),
-            address= {
-                'label': address.get('label')
-            },
+            label=address.get('label'),
+            address=address,
             blocks=build_blocks(es_place, lang, verbosity)
         )
