+++ idunn/api/pois.py
@@ -1,29 +1,12 @@
 from apistar.exceptions import NotFound
 from elasticsearch import Elasticsearch
 
-from idunn.api.place import DEFAULT_VERBOSITY
-from idunn.api.locations.poi import POI
+from idunn.api.place import POI
 from idunn.utils.settings import Settings
-
-def fetch_es_poi(id, es) -> dict:
-    es_pois = es.search(index='munin_poi',
-                        body={
-                            "filter": {
-                                "term": {"_id": id}
-                            }
-                        })
-
-    es_poi = es_pois.get('hits', {}).get('hits', [])
-    if len(es_poi) == 0:
-        raise NotFound(detail={'message': f"poi '{id}' not found"})
-    result = es_poi[0]['_source']
-
-    # Flatten properties into result
-    properties = {p.get('key'): p.get('value') for p in result.get('properties')}
-    result['properties'] = properties
-    return result
+from idunn.api.utils import fetch_es_poi, DEFAULT_VERBOSITY
 
 def get_poi(id, es: Elasticsearch, settings: Settings, lang=None) -> POI:
+    """Handler that returns points-of-interest"""
     if not lang:
         lang = settings['DEFAULT_LANGUAGE']
     lang = lang.lower()
