+++ tests/test_full.py
@@ -47,8 +47,20 @@ def test_full():
             ],
             "type": "Point"
         },
+        "label": "62B Rue de Lille (Paris)",
         "address": {
             "label": "62B Rue de Lille (Paris)"
+        },
+        "address": {
+            "city": "Paris",
+            "citycode": "75056",
+            "housenumber": "62B",
+            "id": "addr:2.326285;48.859635",
+            "label": "62B Rue de Lille (Paris)",
+            "name": "62B Rue de Lille",
+            "postcode": "75007",
+            "street": "Rue de Lille",
+            "type": "house"
         },
         "blocks": [
             {
